fix(pubsublite): close api clients when publisher clients have terminated (#6867)
Refactors the closing of API clients, such that it is done automatically if the PublisherClient permanently fails due to error.

`compositeService` is now passed a `closeable`, which is closed when the `serviceTerminated` state is reached. Users don't necessarily need to call `PublisherClient.Stop` if it has permanently failed in order to close the API clients, though it is good practice to always call `Stop` when a publisher is not longer used.