test(compute): remove TestCapitalLetter (#6908)
This test has been flaky for a long time. If we want something like this test in the future I think it needs to be re-structured to be successful. For now we will remove it as it is not providing value as is and since writing there are now many tests in our samples that should cover these types of use-cases.

Fixes: #5988