fix(bigquery/storage/managedwriter): remove old header routing
Previously, we needed to inject a routing header when opening the
bidi AppendRows connection to successfully route traffic to the
correct region.  Routing no longer needs this explicit hint, and
works by examining the requests on the stream.

This change removes the header injection, and also simplifies the
open contract we use to no longer include a stream ID, which was
in place solely for the injection.

Related internal issue: 185842996