chore(genbot): upgrade base image and protoc (#7032)
Upgrade version of base image to sufficiently new version. This gets us on a newer version of Alpine Linux (`3.16.2`) which gets us on a newer version of `protoc` (`3.18.1`). This is not the newest version of `protoc`, but it includes the necessary `go_package` annotation updates to the common protos that protobuf maintains to upgrade our dependency to the "new" protobuf-go repo distributions.

`docker:20.10`
Alpine Linux: `3.16.2`
Protobuf: `3.18.1-rc3`

Fixes #7031