fix(bigtable/bttest): correct CellsSeenCount and CellsReturnedCount calculations (#7141)
Correct how CellsSeenCount and CellsReturnedCount are calculated in the CBT emulator. The current implementation retrieves the length of the wrong map. It gives incorrect results (compared to production) for certain integration tests when emulator use is allowed.

The emulator and production still disagree on CellsSeenCount for some tests. This could be due to optimizations in production that would be out of scope for the emulator. I'll make the tests a bit more lenient in a follow-up PR.