refactor(bigquery/storage/managedwriter): augment connection pool
This PR adds two references to live code:
* a pool reference on ManagedStream
* a reference to ManagedStream on pendingWrite

The ManagedStream->pool reference is to allow a writer to
resolve where to find its associated connection, retries, lookups, etc.

The reference on the pendingWrite is primarily in service of
retries, particularly when we need to re-enqueue and thus potentially
re-resolve what connection is associated with the writer.  This PR also
moves some of the retry processing code onto the connectionPool in
service to that goal.  As before, this is new code that isn't yet
referenced from existing functionality.

This PR also more substantially starts to carve out connection
management in the pool, providing a basic connection resolver and
eviction capabilities.  This initial implementation is primitive, and
aligns with our current behavior (single unshared connection per
writer).

We also add some testing of the mapping behavior to ensure we're
consistently updating the map for resolution and eviction.

Towards: https://github.com/googleapis/google-cloud-go/issues/7103