refactor: simplify options, instrumentation for multiplexing (#7257)
This PR consolidates where in managedwriter that options apply by moving the destinationTable inside of streamSettings.

This PR also make changes to how instrumentation works.  Currently, we retain a long-lived context on the the ManagedStream and that's used for both connection management and to attach tag context to instrumentation reporting (e.g. to tag metrics with the appropriate keys).  With this change, connection-oriented metrics no longer tag with the writer tags (stream ID, data origin).  We also remove the stream's keys from the default views the package exposes for these metrics.

When we cut over to connections and pools, the context retained on the ManagedStream will only used for metrics tagging, and thus will come from the parent context setup elsewhere (namely the pool context).

Towards: https://togithub.com/googleapis/google-cloud-go/issues/7103