refactor: migrate and and options
This PR does more settings consolidation, and updates/adds new
WriterOption options to propagate settings.

In particular, this PR:
* moves the AppendRows call options into streamSettings
* adds a multiplex flag and option to streamSettings
* adds a call function option into streamSettings

This PR also updates managed stream to use the new option(s) as
appropriate, but most of this is unused here and is in preparation
for a larger cutover of functionality related to the new connection
abstractions.

Towards: https://github.com/googleapis/google-cloud-go/issues/7103