refactor: augment poolRouter interface (and simpleRouter)
This PR extends the poolRouter interface to allow writers to be
registered and removed, and augments the existing simpleRouter to
support the contract.

PR adds a basic test of the router.

A future refactor (when we wire up the new abstractions) will hook
up the functionality properly.

Towards: https://github.com/googleapis/google-cloud-go/issues/7103