refactor: optimize the optimizer
For any reasonable stream, requests on the wire should largely by in the
"optimized" format, with non-essential data redacted.  This PR inverts
behavior of pendingWrite and the optimizers, recording the optimized
form in the pendingWrite.

The change restores previous performance, as for a million sends we're
no longer making proto message copies for 999999 of them.  Instead, we
only manipulate one (the first).