refactor(bigquery/storage/managedwriter): flowcontroller byte reporting (#7581)
This PR allows the flowcontroller to report bytes in flight for flow controllers with a bounded byte definition.  The primary connection load signals for a connection are the inserts/bytes in flight as reported by the flow controller, and this makes the bytes in flight a signal we can use.

Important note: an unbounded flow controller will not report any bytes in flight.  This avoids introducing odd situations due to size normalization where bytes tracked and the actual capacity of the semaphore could get out of sync.

Towards: https://togithub.com/googleapis/google-cloud-go/issues/7103