chore: cleanup mod files (#7691)
Our different tools use too many replace directives for not a lot of benefit. This is step one of two to remove all replaces in gapicgen. This will require a followup commit that up the gensnippet revision to this commit after merging to break this. This was causing gopls to not function well in this module.