chore(storage): move stubs dir to storagepb (#7749)
This is to align this the pattern we use with other apis and will clean up cutting this generation over to owlbot. This also temporarily stops the old generation pipeline for updating storage. A internal CL updating generated paths and another pr to onboard onto owlbot will followup this commit.