test(storage): revert scenario 8 conf test change (#7793)
In #7704 we revised one of the S8 cases to match the upstream
in googleapis/conformance-tests. Unfortunately it seems like that
breaks the test; a download can seemingly recover from 1
broken stream but not 2 in a row. We should investigate and fix
this but for now reverting the change so that the test stops failing.

Updates #7774