chore(storage): migrate all remaining int tests (#7787)
Migrate all remaining integration tests to use multiTransportTest. Most of these skip gRPC and/or JSON but it is good to have all tests migrated with appropriate skips for clarity and consistency.

Found one JSON-related bug in the process, #7786, and another JSON quirk related to error codes.

For gRPC, all of these tests are either not relevant or waiting on backend support (e.g. Notifications and transcoding).