fix(pubsublite): enforce minimum publisher and subscriber timeout of 2 minutes (#7746)
Low timeout values are harmful, as they cause the publisher or subscriber client to terminate if exceeded, which then causes confusion. If a value less than 2 minutes is set, it will be clamped to 2 minutes and a warning will be logged. In the future, a value less than 2 minutes will result in an error.