feat(Default Retention Period): Handle null expiration time in the StartBackupOperation API
Context: Today, cloud customers have to mandatorily specify the expiration timestamp
at the time of create/copy backup. With the default retention period feature, the
expiration timestamp is made optional at the time of create/copy backup. If the
customer doesn't specify the expiration timestamp, the default retention period of
the backup is set to 1 year.

In this commit, making changes to handle null expiration time in the StartBackupOperation
API. Do not set the expiration timestamp in the protobuf if the expire time passed in is
Zero.