fix(bigquery/storage/managedwriter): correct reconnection logic (#8164)
Signalling for an AppendRows stream when schema changes is predicated on
the backend's status for the connection.  For a simplex
(non-multiplexed) connection, the expectation is the client closes and
reconnects to signal there's a change in the schema.

For a connection in multiplex mode, no reconnection is necessary and the
backend will look at the schema for changes.

In managedwriter, we allow a user to specify multiplex at the outset,
but for connections that haven't actually sent writes for more than a
single stream ID the backend doesn't recognize the multiplex status.

This PR expands the interface for send optimizer to signal whether the
optimizer has sent writes for multiple connections, and uses it when
making the determination about schema-based reconnects.  It also
augments the schema evolution test to validate using multiple
combinations of writer and client options.