feat(bigquery/storage/managedwriter): refine connection metrics (#8324)
This PR changes metrics instrumentation in two ways:

The AppendClientOpenView is now tagged with an error dimension, so that
failures to open a connection are clearer.  We use rpc status for the
value, with the expectation that non-rpc errors are tagged as Unknown.

A new metric (and view), AppendRequestReconnectsView provides additional
visibility into when errors during Send trigger reconnection of the
underlying connection.  We attempt to also attribute this to the origin
writer, so it may be tagged by error and source stream.

Towards: https://github.com/googleapis/google-cloud-go/issues/8311