feat(storage): trace span covers life of a Reader (#8390)
Make the tracing span created in NewRangeReader cover the entire lifecycle of the Reader. NewRangeReader returns before the data transfer is complete, so it's more useful to track the entire period until the Reader is closed.

NewRangeReader is still tracked with a span at the level of individual transports (HTTP and gRPC).