feat(auth): add detect package
The purpose of this package is to source credentials via ADC. It
is analogous to the oauth2/google package but is more limited,
currently. It does not provide helper constructors for creating
things like a compute credential and instead has a single
constructor that does it all. Because of this it is more has more
extensible options.

Purposefully missing from this package is support for external
accounts, but some documentation is there. This will come up in a
seperate PR as that has a lot of extra code to review and I tried
to shrink the size of this commit as much as possible.

There are some TODOs in this commit that will be cleaned up in a
future commit that requires more refactoring.