fix(storage): do not inhibit the dead code elimination.
When the go compiler encounters a call to MethodByName(), it
assumes that every method of every reachable type can be accessed,
and nothing may be eliminated.

Go 1.22 has gained support for calls like MethodByName("Foobar").
If the argument to MethodByName() is a compile-time constant, then
the compiler no longer disables the DCE, and only keeps methods
named Foobar.

Rewrite setConditionField() so that MethodByName() is only called with
compile-time constants. This way importing cloud.google.com/go/storage
will not inhibit the DCE.