feat(auth): add httptransport package (#8567)
This package is the equivalent of
google.golang.org/api/transport/http for the new auth library. The
biggest different is the main function does not return an endpoint
so we will want to add a helper in gax or somewhere for this. Also,
there are is an extra helper to add a TokenProvider to a http.Client
which should elevate some user needs for extra extensibility of a
client. They can simply customize away and user the helper to add
auth.