fix(bigquery/storage/managedwriter): automatic retry for multiplex test
In #8599 we caught a flake related to DeadlineExceeded errors.  This
class of error is retriable, so we set the EnableWriteRetries option
when constructing writers to address this.

The DE was likely raised from the service side, as in this case the test
ran for ~32s (deadline is 90) and normal execution time should be
sub-10s.  This is also borne out by the problematic writes, which were
both destined for the same table.

Fixes: https://github.com/googleapis/google-cloud-go/issues/8599