profiler: refactor to avoid (most of the) global state
This change set refactors the profiler to reduce the global state
drastically. It tries to keep the changes minimal on purpose and with
that many opportunities to simplify the code have been left out.

This change set also prepares the profiler to allow for a stop API,
that is once the profiler has been started it can be stopped again and
restarted - even with a different config.

Refs https://github.com/googleapis/google-cloud-go/issues/8632