test(storage): fix flaky RequesterPays
Add retry on 400 and 403 to object deletes

I believe the error was occuring due to the metadata changes not being propagated yet. Other operations are done on the `o` ObjectHandle, which will retry a operation if it encounters a 400 or 403, but these deletes are not retried in that case because they are on a different handle.