chore(vertexai): protoveneer tool and config
Add the tool that is used to generate the "wrappers" for the protos,
and its config for the genai package.

(I prefer to call these types "veneers" because a "wrapper" contains
the thing being wrapped, and these types do not contain the protos. I
decided to resurrect the old name for these clients, because it's an
appropriate metaphor.)