feat(vertexai): use pointers for GenerationConfig fields
Make the fields that are pointers in in the GenerationConfig proto be
pointers in the veneer as well.

This fixes a bug with temperature: when it was a float32, we intpreted zero as nil.
This meant it was impossible to set a zero temperature: attempting to do
so would transmit a nil Temperature to the server, which the server
would interpret as the default. We changed the other fields for consistency.

To make it convenient to set these pointer fields, provide a generic `Ptr` function
as well as a Set method for each field.

Also, remove the model config defaults, which no longer seem to be necessary.