feat(auth): move credentials to base auth package
We are moving the credentials struct to the base auth package as
we intend to use this as the main abstraction for this library.
To better support the new idea of universes having access to the
full credential object will be crutial for the client libraries.
Also, the detect package has been renamed to `credentials` to make
it more generic in the future. We have plans on directly exposing
packages like `externalaccount` in the future so having a better
name here makes sense.

This is the first of two large breaking changes for this module.
This module not in use by our libraries yet, is version 0.1.X and
has always been explictly labeled expirmental in the readme. As
such, we feel confortable making these changes for the long term
health of the module.