feat(auth): move credentials to base auth package (#9590)
We are moving the credentials struct to the base auth package as
we intend to use this as the main abstraction for this library.
To better support the new idea of universes having access to the
full credential object will be crucial for the client libraries.
Also, the detect package has been renamed to `credentials` to make
it more generic in the future. We have plans on directly exposing
packages like `externalaccount` in the future so having a better
name here makes sense.

This is the first of two large breaking changes for this module.
This module not in use by our libraries yet, is version 0.1.X and
has always been explicitly labeled experimental in the readme. As
such, we feel comfortable making these changes for the long term
health of the module.