chore(storage): remove benchmark from go.work
This module is actually depending on 1.22 APIs which tries to pull
dependencies too far forward for other projects. Removing this from
go.work for now. If we lower the Go API level of this module it can
be reintroduced. This will have the effect of you needing to open
VS code from the root of this module to resolve deps correctly.