fix(bigquery/storage/managedwriter): retry improvements
This PR makes two changes to retry behaviors in managedwriter.

In the first, this PR expands the set of conditions that trigger
reconnect when sending the initial request to the backend.

In the second, this PR adds some additional handling for context
cancellations when reading responses back from the service.  In
cases like reconnection, we establish a new Connection, each of
which has it's own associated context.  When draining remaining
writes from a connection that's being shut down, we now pass the
write into a retryer with a status-based error rather than raw
context.Canceled, so we can recover more cleanly if the user is
leveraging write retries.

Related internal issue:
b/326242484