feat(spanner/spannertest): support JSON fields and JSON_VALUE function
This is the minimal level of support required to store values in JSON
columns and retrieve them with the JSON_VALUE function. The JSON_VALUE
function can only be used in query's WHERE clause; that is, you can't
write a query like `SELECT JSON_VALUE(...)`.

This is enough to support generated columns with values derived from
JSON columns, which is a common use case.