chore(aiplatform): move to individual release (#9696)
The `aiplatform` module is a high traffic module wrt changes, and has some volatility we need to manage specifically. This can get in the way of managing the other bulk owned modules, and in times like freezes, make it hard to handle `aiplatform` specific exceptions.

So, we can move it to its own release-please line and manage it independently.