feat(compute/metadata): add context aware functions
This change adds the minimal amount of context aware functionality
so that users can pass a context to all metadata requests. This
does not re-expose all the helper methods this package provides.
We can add context varients for all of these in the future and/or
if we ever create a v2 of this package.

Fixes: #4483