feat(bigquery/storage/managedwriter/adapt): add RANGE support to adapt
This PR adds support for schema conversions and proto conversions to the
adapt subpackage.  For proto conversion, we build a static name for range
message types rather than using the normal first-encountered-message,
as the schema for a range field will also be the consistent for a given
element type.

This PR also moves adapt to using subtests for table-driven testing to
make it easier to isolate specific test cases.