testing(bigquery/storeage/managedwriter): adjust testing boundaries (#10028)
This PR gives a few of the longer-running managedwriter tests a haircut
by reducing iterations.  In many of these cases we're testing for
unusual race conditions and concurrency behaviors, so slightly longer
wall times are to be expected.

Measurement:

On my local cloudtop, `go test -race -short` for the managedwriter package
drops from ~92s to ~42s.