chore(internal/kokoro): filter changes down to unique dir (#10064)
In #9777 there are so many `.go` files being changed that the `internal/kokoro/presubmit.sh` bash scripting was overflowing the conditional check and returning an error. Oddly enough, things would keep going.

We don't really need to check for changes on the `.go` file level if we are just going to run the directory's tests anyways, so let's filter the changes down to unique directory instead.

Ran these commands locally against #9777 and it will fix the bash error we are seeing.