fix(bigquery/storage/managedwriter): faster context failure on send
This is a minor PR that accelerates the failure loop for context
expiration cases.  Previously, the connection abstraction would
deal with expired context on the receiver side, and this change
simply checks context before allowing a new append to succeed.

Tests already existed that exhibited this behavior, but were timing
dependent and masked the issue (the receiver would close the connection
fast enough).

Fixes: https://github.com/googleapis/google-cloud-go/issues/10128