fix(auth): handle non-Transport DefaultTransport (#10162)
Since `http.DefaultTransport` is a `RoundTripper` interface and mutable global variable, it is not safe to assume it is always going to concretely be `*http.Transport`. If it is not, I suppose we should just use the value directly instead of making a `Clone`. The `http.DefaultTransport` being overridden is pretty intentional by application authors, so it is best if we respect that and just reuse it direct.

Fixes #10159