fix(bigtable/bttest): Error when applying a mutation with an empty row key (#9512)
* fix(bigtable/bttest): Error when applying a mutation with an empty row key

Against a real version of Bigtable, an attempt to apply a mutation with
an empty row key leads to the following error being produced

```
unable to apply: rpc error: code = InvalidArgument desc = Row keys must be non-empty
```

bttest does not conform to this behaviour. This commit:

- adds error-handling when given a mutation with an empty row key, and
- adds a regression test.

* chore(bigtable/bttest): Remove a deprecated option from the example

grpc.WithInsecure is deprecated. It was marked as such by v1.43.0 of the
library. See here:

https://github.com/grpc/grpc-go/pull/4718

This commit removes a reference to this from the example comment.

---------

Co-authored-by: Baha Aiman <bahaaiman@google.com>